- Resolved merge conflicts in Cargo.lock, lib-proofs/src/lib.rs, tools/Cargo.toml
- Combined CTO deprecation notices with neural-mesh Plonky2Proof exports
- Preserved all neural-mesh-compression and lib-compression improvements

+++ lib-blockchain/src/contracts/dev_grants/core.rs
@@ -265,7 +265,7 @@ impl DevGrants {
         // Capability-bound: source is derived from ctx, not from parameter
         // ====================================================================
         let burned = token
-            .transfer(ctx, recipient, amt)
+            .transfer(ctx, recipient, amt as u128)
             .map_err(|_| Error::TokenTransferFailed)?;
 
         // ====================================================================
